imporved error handling in the frontend
